backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
psycopg2-binary
python-dotenv
httpx
bcrypt==4.1.2   # <-- Add this line